Revise special case handling for fill-outline-color
The prior code in FillStyleLayer#recalculate handled the case where fill-outline-color was _currently_ undefined, but not the case where it was transitioning from an undefined value to a defined value. In order to handle that case, we must modify the special casing in DataDrivenProperty<T>#interpolate so that it returns a PossiblyEvaluatedPropertyValue containing undefined, rather than a bare undefined value. This makes the return value consistent with the "default value" for fill-outline-color (the one contained in PossiblyEvaluated#_values's prototype), and then both cases can be handled with one conditional in FillStyleLayer#recalculate.